parse_app/requirements.txt
apiclient==1.0.4
asgiref==3.5.2
backports.zoneinfo==0.2.1
bcrypt==3.2.2
beautifulsoup4==4.11.1
bigquery==0.0.21
bs4==0.0.1
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
dacktool==0.0.7
dbstream==0.1.13
Django==4.0.5
django-cors-headers==3.13.0
djangorestframework==3.13.1
google-api-core==2.3.2
google-api-python-client==1.7.11
google-auth==2.8.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.5.2
google-cloud-bigquery==3.2.0
google-cloud-bigquery-storage==2.13.2
google-cloud-core==2.3.1
google-cloud-secret-manager==2.7.2
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.2
googleauthentication==0.0.16
grpc-google-iam-v1==0.12.4
grpcio==1.46.3
grpcio-status==1.46.3
httplib2==0.20.4
idna==3.3
libcst==0.4.4
lxml==4.9.0
mypy-extensions==0.4.3
numpy==1.22.4
oauth2client==4.1.3
oauthlib==3.2.0
packaging==21.3
pandas==1.4.2
paramiko==2.11.0
proto-plus==1.20.6
protobuf==3.20.1
psycopg2-binary==2.9.3
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyNaCl==1.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
sshtunnel==0.1.5
typing-inspect==0.7.1
typing_extensions==4.2.0
uritemplate==3.0.1
urllib3==1.26.9
